remove react dependency completely
