--- src/components/research/D3Graph.astro
@@ -1,27 +1,0 @@
----
-import D3GraphReact, { type DataPoint } from "./D3Graph";
-
-interface Props {
-  title: string;
-  data: DataPoint[];
-  chartType?: "bar" | "pie";
-}
-
-const {
-  title,
-  data,
-  chartType = "bar"
-} = Astro.props;
----
-
-<div class="border-foreground/10 my-10 rounded-2xl border p-6">
-  <h2 class="mb-6 text-2xl font-bold">
-    {title}
-  </h2>
-
-  <D3GraphReact
-    client:load
-    data={data}
-    chartType={chartType}
-  />
-</div>
