+++ src/components/research/ResearchQuote.astro
@@ -9,7 +9,7 @@ const { quote, author } = Astro.props;
 
 <blockquote class="my-8 rounded-2xl border-l-4 border-brandBlue bg-brandBlue/5 p-6 italic">
   <p class="text-foreground text-lg">
-    “{quote}”
+    "{quote}"
   </p>
 
   {author && <footer class="text-foreground/70 mt-4 text-sm font-semibold">— {author}</footer>}
