+++ src/content/research/future-trends-report.mdx
@@ -16,7 +16,7 @@ import DataCard from "@/components/research/DataCard.astro";
 import CalloutBox from "@/components/research/CalloutBox.astro";
 import Timeline from "@/components/research/Timeline.astro";
 import GraphPlaceholder from "@/components/research/GraphPlaceholder.astro";
-import D3Graph from "@/components/research/D3Graph.astro";
+import D3Graph from "@/components//D3Graph.astro";
 
 ## Introduction
